mbps_neurips2026_repro/requirements-adapters.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0
PyYAML>=6.0

# HuggingFace (for DA2, DepthPro)
transformers>=4.35.0

# Depth Anything V3 (custom package — NOT on PyPI)
# Install from source:
#   cd refs/depth-anything-3 && pip install -e .
# Or directly:
#   pip install git+https://github.com/ByteDance-Seed/depth-anything-3.git
# NOTE: Requires Python >=3.9, <=3.13 and CUDA for xformers
depth-anything-3  # installed from refs/depth-anything-3

# For validation metrics
scikit-learn>=1.3.0
